Separated out auth service code.
