+++ Composer/packages/server/src/controllers/auth.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { authService } from '../services/auth';
+import { authService } from '../services/auth/auth';
 import { isElectron } from '../utility/isElectron';
 
 type GetAccessTokenRequest = Request & {
